fix members
