+++ packages/insomnia/src/ui/routes/$organizationId.collaborators.tsx
@@ -1,8 +1,6 @@
-import type { ActionFunction, LoaderFunction } from 'react-router';
+import type { LoaderFunction } from 'react-router';
 
 import { userSession } from '../../models';
-import * as models from '../../models';
-import { invariant } from '../../utils/invariant';
 import { insomniaFetch } from '../insomniaFetch';
 
 interface PaginatedList {
@@ -73,33 +71,5 @@ export const collaboratorsListLoader: LoaderFunction = async ({
     return collaboratorsList;
   } catch (err) {
     return new Error(err.message);
-  }
-};
-
-export const updateMemberRoleAction: ActionFunction = async ({ request, params }) => {
-  const { organizationId, userId } = params;
-
-  invariant(typeof organizationId === 'string', 'Organization ID is required');
-  invariant(typeof userId === 'string', 'User ID is required');
-
-  const formData = await request.formData();
-
-  const roleId = formData.get('roleId');
-  invariant(typeof roleId === 'string', 'Role ID is required');
-
-  try {
-    const user = await models.userSession.getOrCreate();
-    const sessionId = user.id;
-
-    const response = await insomniaFetch<{ enabled: boolean }>({
-      method: 'PATCH',
-      path: `/v1/organizations/${organizationId}/members/${userId}/roles`,
-      data: { roles: [roleId] },
-      sessionId,
-    });
-
-    return response;
-  } catch (err) {
-    throw new Error('Failed to update organization member roles');
   }
 };
